requirements.txt
transformers>=4.40.0
torch>=2.2.0
bitsandbytes>=0.43.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.66.0
sentence-transformers>=2.5.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.4 ; platform_system != "Windows"
